LPS-124465 Avoid unwanted minification of `Liferay` global
As explained in https://issues.liferay.com/browse/LPS-124465 our
minification takes some care to avoid mangling property access:

https://github.com/liferay/liferay-frontend-projects/blob/a33874a2630a07e24cd46421fb346235467ea6c5/projects/npm-tools/packages/npm-scripts/src/config/terser.config.js#L10-L12

Combined with the fact that `Liferay` is usually accessed as a global
and therefore not minified, that means that we can rely on the minifier
leaving strings like `Liferay.Language.get` untouched in the source. The
LanguageFilter can then find those strings and substitute them on the
fly.

In LPS-124465 we see a pattern that breaks this. `Liferay` is passed
into a function, where, as a local binding, the minifier is free to
rewrite it (eg. to `t`), so our source text now reads `t.Language.get`
and the LanguageFilters does not substitute it.

Fix taken here is to avoid that pattern in `portlet.js` and everywhere
else we seem to be doing it (not many places, all found with a crude
regex for `\((\s*\w+\s*,)*\bLiferay\b(\s*,\s*\w+\s*)*\)`; that won't
cover everything, like things split across multiple lines, but it will
catch most things... probably everything).

Instead of:

```js
(function (Liferay) {
    // blah...
})(Liferay);
```

we just do:

```js
(function () {
    // blah...
})();
```

There was one place where we had:

```js
Liferay = window.Liferay || {};
```

Note that the minifier is "smart" enough to know that `Liferay` there is
not a global, and it preserves the name. But prior to this PR, if you
inspect the build output it does the unwanted `t` mangling; ie. before:

    Liferay=window.Liferay||{},function(t){ /* a bunch of stuff */ }(Liferay);

after:

    Liferay=window.Liferay||{},function(){ /* a bunch of stuff */ }();