LPS-76246 Create new RelayStateHelper interface and component to generate RelayState tokens for redirects and retrieve those redirects from the RelayState tokens.
I realize that this probably looks very overengineered. Let me explain
why I designed it this way.

First, we can't "pop" from the Map that is mapping RelayState tokens to
redirects, because there are some cases when we need to retrieve from
the map multiple times, and we don't know at runtime when the last time
is (so we can't simply separate it into "peek" and "pop").

Therefore, the entries will need to stay in the Map, at least for a
while. I imagine that in the most common case, there will only be a
handful of disctinct URLs that users are logging in from, meaning that
the number of distinct redirects will be relatively small. Therefore, I
made the logic bidirectional by including maps in both directions. That
way, we can retrieve the existing RelayState tokens rather than
generating new ones over and over and storing the same URL over and
over, which would be memory-inefficient.

However, it's possible that on some users' systems, they've set things
up in such a way where each user is likely to have their own distinct
redirect URL (for example, maybe there's some userId-related parameter
in the URL that they access the login from). For these cases, it's
important to have the entries expire from the map after a while so that
the size of the map doesn't grow indefinitely. Therefore, I set both
maps to have their entries expire after 10 minutes of having not been
accessed (i.e. 10 mintues after the last read or write operation on the
entry).

Note: On lines 49-51, I included an extra space on purpose, because I
believe order matters:

_redirectsToRelayStateTokens.put(redirect, relayStateToken);

_relayStateTokensToRedirects.put(relayStateToken, redirect);

We want to make sure that the entry in the _relayStateTokensToRedirects
map does not expire before the entry in the _redirectsToRelayStateTokens
map. If it expires after, then it's no big deal (we just need to
generate a new RelayState token in that case), but if it expires before,
it's a huge deal, because then the redirect URL might get lost during
the lookup. So we need to do the write to the
_relayStateTokensToRedirects map after we do the write to the
_redirectsToRelayStateTokens map.